+++ lib/rails_admin/config.rb
@@ -207,7 +207,7 @@ module RailsAdmin
 
       # pool of all found model names from the whole application
       def models_pool
-        excluded = (excluded_models.collect(&:to_s) + %w(RailsAdmin::History PaperTrail::Version PaperTrail::VersionAssociation))
+        excluded = (excluded_models.collect(&:to_s) + %w(RailsAdmin::History PaperTrail::Version PaperTrail::VersionAssociation ActiveStorage::Attachment ActiveStorage::Blob))
 
         (viable_models - excluded).uniq.sort
       end
@@ -302,6 +302,14 @@ module RailsAdmin
       def reset_model(model)
         key = model.is_a?(Class) ? model.name.to_sym : model.to_sym
         @registry.delete(key)
+      end
+
+      # Reset all models configuration
+      # Used to clear all configurations when reloading code in development.
+      # @see RailsAdmin::Engine
+      # @see RailsAdmin::Config.registry
+      def reset_all_models
+        @registry = {}
       end
 
       # Get all models that are configured as visible sorted by their weight and label.