+++ spec/dummy_app/db/migrate/20180701084251_create_active_storage_tables.active_storage.rb
@@ -1,0 +1,30 @@
+# This migration comes from active_storage (originally 20170806125915)
+class CreateActiveStorageTables < MigrationBase
+  def change
+    create_table :active_storage_blobs do |t|
+      t.string   :key,        null: false
+      t.string   :filename,   null: false
+      t.string   :content_type
+      t.text     :metadata
+      if t.respond_to? :bigint
+        t.bigint   :byte_size,  null: false
+      else
+        t.integer  :byte_size,  null: false
+      end
+      t.string   :checksum,   null: false
+      t.datetime :created_at, null: false
+
+      t.index [:key], unique: true
+    end
+
+    create_table :active_storage_attachments do |t|
+      t.string     :name,     null: false
+      t.references :record,   null: false, polymorphic: true, index: false
+      t.references :blob,     null: false
+
+      t.datetime :created_at, null: false
+
+      t.index [:record_type, :record_id, :name, :blob_id], name: "index_active_storage_attachments_uniqueness", unique: true
+    end
+  end
+end