+++ spec/rails_admin/adapters/active_record_spec.rb
@@ -370,6 +370,66 @@ describe 'RailsAdmin::Adapters::ActiveRecord', active_record: true do
         expect(build_statement(:float, ['', '', 'word2'], 'between')).to be_nil
         expect(build_statement(:float, ['', 'word3', 'word4'], 'between')).to be_nil
       end
+
+      it "supports '_blank' operator" do
+        [['_blank', ''], ['', '_blank']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NULL)"])
+          end
+        end
+      end
+
+      it "supports '_present' operator" do
+        [['_present', ''], ['', '_present']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NOT NULL)"])
+          end
+        end
+      end
+
+      it "supports '_null' operator" do
+        [['_null', ''], ['', '_null']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NULL)"])
+          end
+        end
+      end
+
+      it "supports '_not_null' operator" do
+        [['_not_null', ''], ['', '_not_null']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NOT NULL)"])
+          end
+        end
+      end
+
+      it "supports '_empty' operator" do
+        [['_empty', ''], ['', '_empty']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NULL)"])
+          end
+        end
+      end
+
+      it "supports '_not_empty' operator" do
+        [['_not_empty', ''], ['', '_not_empty']].each do |value, operator|
+          aggregate_failures do
+            expect(build_statement(:integer, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:decimal, value, operator)).to eq(["(field IS NOT NULL)"])
+            expect(build_statement(:float, value, operator)).to eq(["(field IS NOT NULL)"])
+          end
+        end
+      end
     end
 
     describe 'date type queries' do