+++ spec/spec_helper.rb
@@ -12,7 +12,7 @@ SimpleCov.formatters = [SimpleCov::Formatter::HTMLFormatter, Coveralls::SimpleCo
 SimpleCov.start do
   add_filter '/spec/'
   add_filter '/vendor/bundle/'
-  minimum_coverage(91.21)
+  minimum_coverage(CI_ORM == :mongoid ? 90.37 : 91.21)
 end
 
 require File.expand_path('../dummy_app/config/environment', __FILE__)
@@ -53,6 +53,7 @@ end
 
 require 'capybara/poltergeist'
 Capybara.javascript_driver = :poltergeist
+Capybara.server = :webrick
 
 RSpec.configure do |config|
   config.expect_with :rspec do |c|