+++ CHANGELOG.md
@@ -2,14 +2,35 @@
 
 ## [Unreleased](https://github.com/sferik/rails_admin/tree/HEAD)
 
-[Full Changelog](https://github.com/sferik/rails_admin/compare/v1.3.0...HEAD)
+[Full Changelog](https://github.com/sferik/rails_admin/compare/v1.4.0...HEAD)
+
+
+## [1.4.0](https://github.com/sferik/rails_admin/tree/v1.4.0) - 2018-07-22
+
+[Full Changelog](https://github.com/sferik/rails_admin/compare/v1.3.0...v1.4.0)
+
+### Added
+- Support for ActiveStorage([#2990](https://github.com/sferik/rails_admin/issues/2990), [#3037](https://github.com/sferik/rails_admin/pull/3037))
+- Support for multiple file upload for ActiveStorage and CarrierWave ([5bb2d375](https://github.com/sferik/rails_admin/commit/5bb2d375a236268e51c7e8682c2d110d9e52970f))
+- Support for Mongoid 7.0([9ef623f6](https://github.com/sferik/rails_admin/commit/9ef623f6cba73adbf86833d9eb07f1be3924a133), [#3013](https://github.com/sferik/rails_admin/issues/3013))
+- Support for CanCanCan 2.0([a32d49e4](https://github.com/sferik/rails_admin/commit/a32d49e4b96944905443588a1216b3362ee64c1a), [#2901](https://github.com/sferik/rails_admin/issues/2901))
+- Support for Pundit 2.0([bc60c978](https://github.com/sferik/rails_admin/commit/bc60c978adfebe09cdad2c199878d8ff966374f1))
+- Support for jquery-ui-rails 6.0([#2951](https://github.com/sferik/rails_admin/issues/2951), [#3003](https://github.com/sferik/rails_admin/issues/3003))
+
+### Fixed
+- Make code reloading work([#3041](https://github.com/sferik/rails_admin/pull/3041))
+- Improved support for Rails API mode, requiring needed middlewares in engine's initializer([#2919](https://github.com/sferik/rails_admin/issues/2919), [#3006](https://github.com/sferik/rails_admin/pull/3006))
+- Make the link text to uploaded file shorter, instead of showing full url([#2983](https://github.com/sferik/rails_admin/pull/2983))
+- Fix duplication of filters on browser back([#2998](https://github.com/sferik/rails_admin/pull/2998))
+- Fix "can't modify frozen array" exception on code reload([#2999](https://github.com/sferik/rails_admin/pull/2999))
+- Fix incorrectly comparing numeric columns with empty string when handling blank operator([#3007](https://github.com/sferik/rails_admin/pull/3007))
 
 
 ## [1.3.0](https://github.com/sferik/rails_admin/tree/v1.3.0) - 2018-02-18
 
 [Full Changelog](https://github.com/sferik/rails_admin/compare/v1.2.0...v1.3.0)
 
-#### Added
+### Added
 - Configurability for forgery protection setting([#2989](https://github.com/sferik/rails_admin/pull/2989))
 - Configurability for the number of audit records displayed into dashboard([#2982](https://github.com/sferik/rails_admin/pull/2982))
 - Add limited pagination mode, which doesn't require count query([#2968](https://github.com/sferik/rails_admin/pull/2968))
@@ -33,7 +54,7 @@
 
 [Full Changelog](https://github.com/sferik/rails_admin/compare/v1.1.1...v1.2.0)
 
-#### Added
+### Added
 - Add ILIKE support for PostgreSQL/PostGIS adapter, multibyte downcase for other adapters([#2766](https://github.com/sferik/rails_admin/pull/2766))
 - Support for UUID query([#2766](https://github.com/sferik/rails_admin/pull/2766))
 - Support for Haml 5([#2840](https://github.com/sferik/rails_admin/pull/2840), [#2870](https://github.com/sferik/rails_admin/pull/2870), [#2877](https://github.com/sferik/rails_admin/pull/2877))